api/leagues-api/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
alchemical==0.5.1
    # via -r requirements.in
alembic==1.7.7
    # via flask-migrate
apifairy==0.9.1
    # via -r requirements.in
apispec==5.1.1
    # via apifairy
blinker==1.4
    # via flask-mail
cachecontrol==0.12.11
    # via firebase-admin
cachetools==5.2.0
    # via google-auth
certifi==2022.9.24
    # via requests
charset-normalizer==2.1.1
    # via requests
click==8.1.0
    # via flask
faker==13.3.4
    # via -r requirements.in
firebase-admin==5.4.0
    # via -r requirements.in
flask==2.1.0
    # via
    #   -r requirements.in
    #   apifairy
    #   flask-cors
    #   flask-httpauth
    #   flask-mail
    #   flask-marshmallow
    #   flask-migrate
    #   flask-sqlalchemy
flask-cors==3.0.10
    # via -r requirements.in
flask-httpauth==4.5.0
    # via
    #   -r requirements.in
    #   apifairy
flask-mail==0.9.1
    # via -r requirements.in
flask-marshmallow==0.14.0
    # via
    #   -r requirements.in
    #   apifairy
flask-migrate==3.1.0
    # via -r requirements.in
flask-sqlalchemy==2.5.1
    # via flask-migrate
google-api-core[grpc]==2.10.1
    # via
    #   firebase-admin
    #   google-api-python-client
    #   google-cloud-core
    #   google-cloud-firestore
    #   google-cloud-storage
google-api-python-client==2.63.0
    # via firebase-admin
google-auth==2.12.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-cloud-core
    #   google-cloud-storage
google-auth-httplib2==0.1.0
    # via google-api-python-client
google-cloud-core==2.3.2
    # via
    #   google-cloud-firestore
    #   google-cloud-storage
google-cloud-firestore==2.7.0
    # via firebase-admin
google-cloud-storage==2.5.0
    # via firebase-admin
google-crc32c==1.5.0
    # via google-resumable-media
google-resumable-media==2.4.0
    # via google-cloud-storage
googleapis-common-protos==1.56.4
    # via
    #   google-api-core
    #   grpcio-status
greenlet==1.1.2
    # via sqlalchemy
grpcio==1.49.1
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.49.1
    # via google-api-core
gunicorn==20.1.0
    # via -r requirements.in
httplib2==0.20.4
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.4
    # via requests
importlib-metadata==4.12.0
    # via flask
itsdangerous==2.1.2
    # via flask
jinja2==3.1.1
    # via flask
mako==1.2.0
    # via alembic
markupsafe==2.1.1
    # via
    #   jinja2
    #   mako
marshmallow==3.15.0
    # via
    #   flask-marshmallow
    #   marshmallow-sqlalchemy
    #   webargs
marshmallow-sqlalchemy==0.28.0
    # via -r requirements.in
msgpack==1.0.4
    # via cachecontrol
packaging==21.3
    # via
    #   marshmallow
    #   webargs
proto-plus==1.22.1
    # via google-cloud-firestore
protobuf==4.21.7
    # via
    #   google-api-core
    #   google-cloud-firestore
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pyjwt==2.3.0
    # via -r requirements.in
pyparsing==3.0.7
    # via
    #   httplib2
    #   packaging
python-dateutil==2.8.2
    # via faker
python-dotenv==0.20.0
    # via -r requirements.in
requests==2.28.1
    # via
    #   cachecontrol
    #   google-api-core
    #   google-cloud-storage
rsa==4.9
    # via google-auth
six==1.16.0
    # via
    #   flask-cors
    #   flask-marshmallow
    #   google-auth
    #   google-auth-httplib2
    #   grpcio
    #   python-dateutil
sqlalchemy==1.4.32
    # via
    #   alchemical
    #   alembic
    #   flask-sqlalchemy
    #   marshmallow-sqlalchemy
uritemplate==4.1.1
    # via google-api-python-client
urllib3==1.26.12
    # via requests
webargs==8.1.0
    # via apifairy
werkzeug==2.1.0
    # via flask
zipp==3.8.1
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
